feature/texoit-project
Created new project to TexoIT test.

This project consists of a front-end application developed with Ionic, Angular and TypeScript. The goal is to provide an interface to visualize data on the nominees and winners of the Golden Raspberry Awards "Worst Film" category, based on data provided by an external API.